+++ 1.20.4/Common/src/main/java/fuzs/puzzleslib/api/container/v1/ContainerImpl.java
@@ -82,7 +82,7 @@ public interface ContainerImpl extends Container {
 
     @Override
     default void setChanged() {
-
+        // NO-OP
     }
 
     @Override